TST: move tests to io/pytables/test_compat.py
